pkg/scorecard: move pkg/scorecard/lib to pkg/scorecard
