+++ internal/pkg/scorecard/helpers.go
@@ -14,7 +14,7 @@
 
 package scorecard
 
-import sclib "github.com/operator-framework/operator-sdk/pkg/scorecard/lib"
+import sclib "github.com/operator-framework/operator-sdk/pkg/scorecard"
 
 // These functions should be in the public test definitions file, but they are not complete/stable,
 // so we'll keep these here until they get fully implemented
