+++ internal/pkg/scorecard/olm_tests.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 
 	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
-	sclib "github.com/operator-framework/operator-sdk/pkg/scorecard/lib"
+	sclib "github.com/operator-framework/operator-sdk/pkg/scorecard"
 
 	olmapiv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
 	log "github.com/sirupsen/logrus"
